net: refactor vnet_hdr methods
We're going to use it from different modules, so move them upper in the
crate.
